+++ src/Compatibility/Core/src/WinUI/VisualElementRenderer.cs
@@ -750,7 +750,6 @@ namespace Microsoft.Maui.Controls.Compatibility.Platform.UWP
 			if (_tracker == null)
 				return;
 
-			_tracker.PreventGestureBubbling = PreventGestureBubbling;
 			_tracker.Control = Control;
 			_tracker.Element = Element;
 			_tracker.Container = ContainerElement;
