+++ src/Controls/samples/Controls.Sample/ViewModels/GesturesViewModel.cs
@@ -17,11 +17,8 @@ namespace Maui.Controls.Sample.ViewModels
 				"Pan Gesture."),
 			new SectionModel(typeof(SwipeGestureGalleryPage), "Swipe Gesture",
 				"Swipe Gesture."),
-
-#if WINDOWS || MACCATALYST
-			new SectionModel(typeof(ClickGestureGalleryPage), "Click Gesture",
+			new SectionModel(typeof(TapGestureGalleryPage), "Click Gesture",
 				"Click Gesture."),
-#endif
 		};
 	}
 }
